requirements.txt
fastapi
uvicorn[standard]
openai-whisper
torch
torchaudio
soundfile
requests
python-multipart
omegaconf
